refactor: move Separate to Enum namespace
